+++ actionview/CHANGELOG.md
@@ -1,3 +1,26 @@
+*   `tag.attributes` returns `Hash`-like `ActionView::Attributes` instance
+
+    When called outside of a rendering context, <tt>tag.attributes</tt>
+    will return a <tt>Hash</tt>-like object that knows how to render
+    itself to HTML:
+
+        primary = { class: "bg-red-500 text-white" }
+        large = { class: "text-lg p-4" }
+
+        tag.attributes(primary, large).to_h
+        # => { class: "bg-red-500 text-white text-lg p-4" }
+
+        tag.attributes(primary, large).to_s
+        # => "class=\"bg-red-500 text-white text-lg p-4\""
+
+        button_tag "Click me!", tag.attributes(primary, large)
+        # => <button name="button" type="submit" class="bg-red-500 text-white text-lg p-4">Click me!</button>
+
+        tag.button "Click me!", id: "cta", **tag.attributes(primary, large)
+        # => <button id="cta" class="bg-red-500 text-white text-lg p-4">Click me!</button>
+
+    *Sean Doyle*
+
 *   `tag.attributes` accepts a variable number of `Hash` arguments, then merges
     them from left to right: